fix(exec): make signal exit handling platform-aware
