+++ api/service/exec/exit.go
@@ -4,8 +4,6 @@ package exec
 
 import (
 	"errors"
-	osexec "os/exec"
-	"syscall"
 )
 
 // ExitStatus describes how a child process finished.
@@ -64,19 +62,4 @@ func WaitFor(p Process) ExitStatus {
 		return reporter.AwaitExit()
 	}
 	return ClassifyExit(p.Wait())
-}
-
-// exitSignal reads the signal that killed the child from the platform wait
-// status. Executors that report only an exit code carry no wait status and
-// report no signal.
-func exitSignal(err error) int {
-	var exitErr *osexec.ExitError
-	if !errors.As(err, &exitErr) || exitErr.ProcessState == nil {
-		return 0
-	}
-	status, ok := exitErr.Sys().(syscall.WaitStatus)
-	if !ok || !status.Signaled() {
-		return 0
-	}
-	return int(status.Signal())
 }
