+++ api/service/exec/exit_test.go
@@ -6,7 +6,6 @@ import (
 	"errors"
 	"io"
 	osexec "os/exec"
-	"syscall"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -36,23 +35,6 @@ func TestClassifyExitKeepsAnErrorThatIsNotAnExit(t *testing.T) {
 
 	require.ErrorIs(t, status.Err, wait)
 	require.Zero(t, status.Code)
-}
-
-// A child killed by a signal has no exit code of its own. It is reported as
-// 128+signal, with the signal itself alongside.
-func TestClassifyExitReportsSignalDeath(t *testing.T) {
-	if _, err := osexec.LookPath("sleep"); err != nil {
-		t.Skip("sleep not available")
-	}
-	cmd := osexec.CommandContext(t.Context(), "sleep", "30")
-	require.NoError(t, cmd.Start())
-	require.NoError(t, cmd.Process.Signal(syscall.SIGKILL))
-
-	status := ClassifyExit(cmd.Wait())
-
-	require.Equal(t, int(syscall.SIGKILL), status.Signal)
-	require.Equal(t, 128+int(syscall.SIGKILL), status.Code)
-	require.NoError(t, status.Err)
 }
 
 func TestClassifyExitReportsOrdinaryExitCode(t *testing.T) {
